refactor: extract hotkey functions to src/input/hotkeys.rs
- Move install_hotkeys, uninstall_hotkeys, reinstall_hotkeys to input module
- Make handler a parameter for better separation of concerns
- Update all call sites to pass hotkey_event_handler
- Remove duplicate functions from main.rs (-89 lines)
